git: verify-commit
